refactor: move fns from commons into ingest script
